Rename data_connection directory to dataProvider. And move component directory to inside the pages
