+++ package.json
@@ -12,6 +12,7 @@
     "ra-data-json-server": "^3.19.0",
     "ra-input-rich-text": "^3.19.0",
     "react": "^16.5.2",
+    "react-ace": "^9.5.0",
     "react-admin": "^3.19.0",
     "react-bootstrap": "^0.32.4",
     "react-dom": "^16.5.2",
